requirements.txt
sqlalchemy==2.0.20
fastapi==0.101.1
uvicorn==0.25.0
pydantic==2.3.0
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.114.0
requests~=2.31.0
h11~=0.14.0
wheel~=0.41.2
cryptography~=41.0.7
rsa~=4.9
pyasn1~=0.5.1
setuptools~=68.2.0
ecdsa~=0.18.0
typing_extensions~=4.9.0
sniffio~=1.3.0
six~=1.16.0
starlette~=0.27.0
pydantic_core~=2.6.3
anyio~=4.2.0
click~=8.1.7
pycparser~=2.21
bcrypt~=4.1.2
googleapiclient~=2.24.0